refactor: top up dialog to DDD
